requirements.txt
bleach==6.0.0
branca==0.6.0
certifi==2023.5.7
cffi==1.15.0
charset-normalizer==3.1.0
click==8.0.1
cryptography==37.0.2
Deprecated==1.2.14
Flask==2.0.1
folium==0.14.0
idna==3.4
importlib-resources==6.0.0
itsdangerous==2.0.1
Jinja2==3.0.1
limits==3.5.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
numpy==1.24.3
ordered-set==4.1.0
packaging==23.1
peewee==3.14.10
pycparser==2.21
Pygments==2.15.1
PyMySQL==1.0.2
python-dotenv==0.17.1
regex==2023.6.3
requests==2.31.0
rich==13.4.2
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.3
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.15.0
